pyproject.toml
[tool.poetry]
name = "notification-service"
version = "0.1.0"
description = ""
authors = ["Sergey Yalovoy <s.yalovoy@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
sqlalchemy = "^2.0.31"
pydantic = "^2.7.4"
alembic = "^1.13.1"
asyncpg = "^0.29.0"
uvicorn = "^0.30.1"
pydantic-settings = "^2.3.4"
python-dotenv = "^1.0.1"
psycopg2 = "^2.9.9"
fastapi-filter = {extras = ["sqlalchemy"], version = "^2.0.0"}
fastapi-pagination = "^0.12.26"
pywebpush = "^2.0.0"
firebase-admin = "^6.5.0"
aio-pika = "^9.4.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
